moved cmds back for consistency
