+++ cli/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/NethermindEth/juno/cli/cmd"
+
+func main() {
+	cmd.Execute()
+}
