--- cmd/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/NethermindEth/juno/pkg/cmd"
-
-func main() {
-	cmd.Execute()
-}
