--- pkg/cmd/exampleValues.txt
@@ -1,5 +1,0 @@
-Example tx:
-
-> Probably sent something
-0x21601cf55af436dfb73178347973c57cbedc80ff442f6d841815446e8947a94
-
