--- pkg/cmd/long.txt
@@ -1,8 +1,0 @@
-       _                    
-      | |                   
-      | |_   _ _ __   ___   
-  _   | | | | | '_ \ / _ \  
- | |__| | |_| | | | | (_) | 
-  \____/ \__,_|_| |_|\___/  
-
-Juno is a Go implementation of a StarkNet full node client made with ❤️ by Nethermind.
