--- pkg/cmd/tests/test.cairo
@@ -1,6 +1,0 @@
-func main():
-    [ap] = 1000; ap++
-    [ap] = 2000; ap++
-    [ap] = [ap - 2] + [ap - 1]; ap++
-    ret
-end
